+++ test/protractor/spec/lforms_fhirdata.spec.js
@@ -601,6 +601,7 @@ for (var i=0, len=fhirVersions.length; i<len; ++i) {
 
           it('should merge FHIR SDC QuestionnaireResponse data back into the form', function() {
             tp.openUSSGFHTVertical();
+            tp.setFHIRVersion(fhirVersion);
 
             element(by.id("merge-qr")).click();
 
@@ -637,13 +638,15 @@ for (var i=0, len=fhirVersions.length; i<len; ++i) {
 
         describe('Converted Questionnaire', function () {
           beforeEach(function() {
-            tp.openQuestionnaire();
+            tp.openBaseTestPage();
+            tp.setFHIRVersion(fhirVersion);
+            tp.loadFromTestData('4712701.json', fhirVersion);
           });
 
           it('should be able to show a converted questionnaire', function() {
-             // Check to see that the last question has rendered
-             expect(element(by.id('label-/4/TBD3/TBD19/TBD20/1/1/1/1')).getText()).toBe(
-               "Rezidiv/Progress aufgetreten");
+            // Check to see that the last question has rendered
+            expect(element(by.id('label-/4/TBD3/TBD19/TBD20/1/1/1/1')).getText()).toBe(
+              "Rezidiv/Progress aufgetreten");
           });
 
           it('should have functioning skiplogic when codes are not present', function() {